Macro spella::constant

macro_rules! constant {
    () => { ... };
    ($(#[$attr:meta])* $vis:vis const $Name:ident: $T:ty = $val:expr; $($rest:tt)*) => { ... };
    ($(#[$attr:meta])* $vis:vis static $Name:ident: $T:ty = $val:expr; $($rest:tt)*) => { ... };
}

Creates a marker type that represents a constant value.

Syntax

constant! {
  // Compile-time constant value
  const NAME: TYPE = EXPR;

  // Runtime constant value
  static NAME: TYPE = EXPR;
}

If it contains static's one, it works only in functions.

How it works

This macro defines an empty enum type that implements Constant<TYPE> with an accosiated function fn get() -> TYPE which returns the given EXPR.

const

EXPR will simply be declared as a normal Rust's const value in a private scope.

static

It declares a thread local storage in a private scope, and immediately sets EXPR to that.

Examples

fn mod_mul<Mod: Constant<u32>>(x: u32, y: u32) -> u32 {
  x * y % Mod::get()
}

fn main() {
  let x = 10;

  constant! {
    const MOD1: u32 = 2;
    static MOD2: u32 = x;
  }

  assert_eq!(mod_mul::<MOD1>(7, 3), 7 * 3 % 2);
  assert_eq!(mod_mul::<MOD2>(7, 3), 7 * 3 % x);
}